// frontend/vendor/tradingview-pro/src/types/trading.ts
export interface IndicatorConfig {
  id: string;
  name: string;
  shortName: string;
  type: 'overlay' | 'pane';
  visible: boolean;
  color: string;
  color2?: string;
  color3?: string;
  params: Record<string, number | string | boolean>;
  values?: number[] | { [key: string]: number[] };
}